requirements.txt
# Core
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0

# Email
secure-smtplib>=0.1.0

# Data
pydantic>=2.0.0
jinja2>=3.1.0

# Scheduling (optional, for local runs)
schedule>=1.2.0
python-dotenv>=1.0.0
